mkos-knowledge-immune-system/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "akm"
version = "0.1.0"
description = "AI Knowledge Management - Cross-project knowledge search"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5",
    "anthropic>=0.80.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.0",
]

[project.scripts]
akm = "akm.cli:main"
